extract file-downloader crate
