+++ download-utils/Cargo.toml
@@ -10,10 +10,8 @@ license = { workspace = true }
 edition = { workspace = true }
 
 [dependencies]
-console = { workspace = true }
-indicatif = { workspace = true }
 log = { workspace = true }
-reqwest = { workspace = true, features = ["blocking", "brotli", "deflate", "gzip", "rustls-tls", "json"] }
+solana-file-downloader = { workspace = true }
 solana-runtime = { workspace = true }
 solana-sdk = { workspace = true }
